refactor(dnd5e): update imports and remove lru_cache decorators
Update import paths to use absolute imports (src.domain...) instead of relative ones
Remove lru_cache decorators from API client methods to prevent potential memory issues
Add optional level field to ResultsTemplate dataclass
