+++ src/domain/entities/api_data/dnd5e/dnd5e.py
@@ -1,4 +1,5 @@
 from dataclasses import dataclass
+from typing import Optional
 
 
 @dataclass
@@ -6,3 +7,4 @@ class ResultsTemplate:
     index: str
     name: str
     url: str
+    level: Optional[int] = None
