+++ src/domain/entities/api_data/dnd5e/spells.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass
 from typing import List
 
-from domain.entities.api_data.dnd5e.dnd5e import ResultsTemplate
+from src.domain.entities.api_data.dnd5e.dnd5e import ResultsTemplate
 
 
 @dataclass
